+++ core/search/scorable.go
@@ -5,7 +5,7 @@ package search
 // 允许访问查询的分数
 type Scorable interface {
 	// Score Returns the Score of the current document matching the query.
-	Score() (float32, error)
+	Score() (float64, error)
 
 	// SmoothingScore Returns the smoothing Score of the current document matching the query. This Score
 	// is used when the query/term does not appear in the document, and behaves like an idf. The smoothing
@@ -15,7 +15,7 @@ type Scorable interface {
 	// Smoothing scores are described in many papers, including: Metzler, D. and Croft, W. B. , "Combining
 	// the Language Model and Inference Network Approaches to Retrieval," Information Processing and Management
 	// Special Issue on Bayesian Networks and Information Retrieval, 40(5), pp.735-750.
-	SmoothingScore(docId int) (float32, error)
+	SmoothingScore(docId int) (float64, error)
 
 	// DocID Returns the doc ID that is currently being scored.
 	DocID() int
@@ -24,7 +24,7 @@ type Scorable interface {
 	// documents whose Score is less than the given minScore. This is a no-op by default. This method
 	// may only be called from collectors that use ScoreMode.TOP_SCORES, and successive calls may
 	// only set increasing values of minScore.
-	SetMinCompetitiveScore(minScore float32) error
+	SetMinCompetitiveScore(minScore float64) error
 
 	// GetChildren Returns child sub-scorers positioned on the current document
 	GetChildren() ([]ChildScorable, error)
@@ -33,11 +33,11 @@ type Scorable interface {
 type ScorableDefault struct {
 }
 
-func (*ScorableDefault) SmoothingScore(docId int) (float32, error) {
+func (*ScorableDefault) SmoothingScore(docId int) (float64, error) {
 	return 0, nil
 }
 
-func (*ScorableDefault) SetMinCompetitiveScore(minScore float32) error {
+func (*ScorableDefault) SetMinCompetitiveScore(minScore float64) error {
 	return nil
 }
